initial webcam support.
This doesn't realy talk to v4l yet but implements a simple http endpoint and the subprocess handling.
The idea here is to use a subprocess (of ourself) to handle all v4l communication to avoid issues with the process in case of kernel / driver issues:
If the subprocess gets stuck, reads would just time out instead of bringing the whole process down (possibling messing up the print).